# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Course Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a single course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /courses/{courseId}/videos:
    get:
      operationId: listCourseVideos
      tags: [videos]
      summary: List video sections for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoSection"
  /courses/{courseId}/files:
    get:
      operationId: listCourseFiles
      tags: [files]
      summary: List file categories for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileCategory"
  /courses/{courseId}/sessions:
    get:
      operationId: listCourseSessions
      tags: [sessions]
      summary: List interactive sessions for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InteractiveSession"
  /courses/{courseId}/contact:
    get:
      operationId: getCourseContact
      tags: [courses]
      summary: Get course contact info
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseContact"
  /progress/video/{videoId}:
    post:
      operationId: markVideoWatched
      tags: [progress]
      summary: Mark a video as watched
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /progress/course/{courseId}:
    get:
      operationId: getCourseProgress
      tags: [progress]
      summary: Get progress for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseProgress"
...
  /devices/{id}:
    delete:
      operationId: removeDevice
      tags: [devices]
      summary: Logout from a specific device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        approved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - approved
        - createdAt
...
    Teacher:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
...
    CourseDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        coverImage:
          type: string
        status:
          type: string
        createdAt:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        instructorName:
          type: ["string", "null"]
        teacherId:
          type: ["integer", "null"]
        enrolled:
          type: boolean
      required:
        - id
        - title
        - description
        - coverImage
        - status
        - createdAt
        - enrolled
...
    VideoSection:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sortOrder:
          type: integer
        courseId:
          type: integer
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
      required:
        - id
        - title
        - sortOrder
        - courseId
        - videos
    Video:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string
        sortOrder:
          type: integer
        sectionId:
          type: integer
        watched:
          type: boolean
      required:
        - id
        - title
        - url
        - sortOrder
        - sectionId
    FileCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        courseId:
          type: integer
        files:
          type: array
          items:
            $ref: "#/components/schemas/CourseFile"
      required:
        - id
        - name
        - courseId
        - files
    CourseFile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        categoryId:
          type: integer
      required:
        - id
        - name
        - url
        - categoryId
    InteractiveSession:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        link:
          type: string
        scheduledAt:
          type: ["string", "null"]
        courseId:
          type: integer
      required:
        - id
        - title
        - link
        - courseId
    CourseContact:
      type: object
      properties:
        instructorName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        teacher:
          $ref: "#/components/schemas/Teacher"
    CourseProgress:
      type: object
      properties:
        courseId:
          type: integer
        totalVideos:
          type: integer
        watchedVideos:
          type: integer
        progress:
          type: number
        watchedVideoIds:
          type: array
          items:
            type: integer
      required:
        - courseId
        - totalVideos
        - watchedVideos
        - progress
        - watchedVideoIds
...