# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Course Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}