# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Course Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: number;
  title: string;
  url: string;
  sortOrder: number;
  sectionId: number;
  watched?: boolean;
}